add user deletion

+++ src/YLunch.Application/Services/UserService.cs
@@ -1,6 +1,7 @@
 using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
+using YLunch.Application.Exceptions;
 using YLunch.Domain.DTO.UserModels;
 using YLunch.Domain.ModelsAggregate.UserAggregate;
 using YLunch.Domain.Services.Database.Repositories;
@@ -31,9 +32,13 @@ namespace YLunch.Application.Services
             return new UserAsCustomerDetailsReadDto(user);
         }
 
-        public async Task DeleteUserById(string id)
+        public async Task DeleteUserByUsername(string username)
         {
-            var user = await _userRepository.GetById(id);
+            var user = await _userRepository.GetFullUser(username);
+            if (user is null)
+            {
+                throw new NotFoundException($"User with username: '{username}' not found");
+            }
             await _userRepository.Delete(user);
         }
     }
